rename: cookbook list to apps
