+++ src/tapis-ui/components/apps/index.ts
@@ -1,0 +1,1 @@
+export { default as FileListing } from './AppListing';
