--- src/tapis-ui/components/cookbooks/index.ts
@@ -1,1 +1,0 @@
-export { default as FileListing } from './CookbookListing';
